turnovers set up
For this, I borrowed code above to build a turnovers created and given up for both home and away Then I build the structure for the differences and the difference differential. How I have it built, if the turnover difference differential is negative, that means a team have given up more than they have created and vice vera. So in this instance, a high positive number here would indicate a team that is successful at creating turnovers as well as not giving them up.

My issue I have now is figuring out what variables I want to use to create this variable, but also have questions about where each variable should be included or what it truly means. Going to give it some more thought, and then revist tomorrow.